[#1100] restructuring plot_diagnostic_geometric_coverage_slice()
