+++ tofu/data/_class8_plot_coverage_slice.py
@@ -7,16 +7,12 @@
 
 # Common
 import numpy as np
-import matplotlib.pyplot as plt
-import matplotlib.gridspec as gridspec
-from mpl_toolkits.mplot3d import Axes3D
 import datastock as ds
 
 
-# tofu
-from ..geom import _comp_solidangles
-from . import _class8_plot
-from . import _generic_check
+# local
+from . import _class8_plot_coverage_slice_broadband as _broadband
+from . import _class8_plot_coverage_slice_spectro as _spectro
 
 
 # ###############################################################
@@ -44,7 +40,9 @@ def main(
     DR=None,
     DZ=None,
     Dphi=None,
+    # solid angle
     config=None,
+    visibility=None,
     # solid angle
     n0=None,
     n1=None,
@@ -59,14 +57,7 @@ def main(
     plot_config=None,
     fs=None,
     dmargin=None,
-    vmin_cam0=None,
-    vmax_cam0=None,
-    vmin_cam=None,
-    vmax_cam=None,
-    vmin_cam_lamb=None,
-    vmax_cam_lamb=None,
-    vmin_plane=None,
-    vmax_plane=None,
+    dvminmax=None,
 ):
 
     # ------------
@@ -79,6 +70,7 @@ def main(
         lop_pre, lop_post,
         res, margin_par, margin_perp,
         vect, segment, phi, Z, key_mesh,
+        visibility,
         verb, plot,
         indref, indplot,
         plot_config,
@@ -101,13 +93,15 @@ def main(
         DR=DR,
         DZ=DZ,
         Dphi=Dphi,
+        # solid angle
+        config=config,
+        visibility=visibility,
         # bool
         verb=verb,
         plot=plot,
         # plotting
         indplot=indplot,
         plot_config=plot_config,
-        config=config,
     )
 
     # -----------------
@@ -144,7 +138,7 @@ def main(
 
         dpts = coll.get_sample_mesh_3d_slice(
             key=key_mesh,
-            res=res,
+            res=res[0],
             phi=phi,
             Z=Z,
             DR=DR,
@@ -165,40 +159,48 @@ def main(
         x0, x1 = None, None
         dS = None
 
+    # ---------------------------
+    # get compute & plot function
+    # ---------------------------
+
+    if spectro is True:
+        _compute = _spectro._compute
+        if vect is None:
+            _plot = _spectro._plot_from_mesh
+        else:
+            _plot = _spectro._plot_from_los
+
+    else:
+        _compute = _broadband._compute
+        if vect is None:
+            _plot = _broadband._plot_from_mesh
+        else:
+            _plot = _broadband._plot_from_los
+
     # ----------
     # compute
     # ----------
 
-    if spectro:
-        dout = _spectro(
-            coll=coll,
-            key_diag=key_diag,
-            key_cam=key_cam,
-            # pts
-            ptsx=ptsx,
-            ptsy=ptsy,
-            ptsz=ptsz,
-            # solid angle
-            n0=n0,
-            n1=n1,
-            # res
-            res_lamb=res_lamb,
-        )
-
-    else:
-        dout = _nonspectro(
-            coll=coll,
-            key_cam=key_cam,
-            doptics=doptics,
-            par=parallel,
-            is2d=is2d,
-            # points
-            # pts
-            ptsx=ptsx,
-            ptsy=ptsy,
-            ptsz=ptsz,
-            config=config,
-        )
+    dout = _compute(
+        coll=coll,
+        key_diag=key_diag,
+        key_cam=key_cam,
+        doptics=doptics,
+        par=parallel,
+        is2d=is2d,
+        # pts
+        ptsx=ptsx,
+        ptsy=ptsy,
+        ptsz=ptsz,
+        # solid angle
+        n0=n0,
+        n1=n1,
+        # res
+        res_lamb=res_lamb,
+        # config
+        config=config,
+        visibility=visibility,
+    )
 
     # -------------
     # format output
@@ -223,6 +225,9 @@ def main(
         'x1': x1,
         'dS': dS,
         'vect': vect,
+        'Z': Z,
+        'Z': Z,
+        'phi': phi,
     })
 
     # -------
@@ -230,47 +235,18 @@ def main(
     # -------
 
     if plot is True:
-        if vect is not None:
-            _plot_from_los(
-                coll=coll,
-                # extra
-                indplot=indplot,
-                dax=dax,
-                plot_config=plot_config,
-                fs=fs,
-                dmargin=dmargin,
-                vmin_cam0=vmin_cam0,
-                vmax_cam0=vmax_cam0,
-                vmin_cam=vmin_cam,
-                vmax_cam=vmax_cam,
-                vmin_cam_lamb=vmin_cam_lamb,
-                vmax_cam_lamb=vmax_cam_lamb,
-                vmin_plane=vmin_plane,
-                vmax_plane=vmax_plane,
-                # dout
-                **dout,
-            )
-
-        elif phi is not None:
-            _plot_from_cross(
-                coll=coll,
-                # extra
-                indplot=indplot,
-                dax=dax,
-                plot_config=plot_config,
-                fs=fs,
-                dmargin=dmargin,
-                vmin_cam0=vmin_cam0,
-                vmax_cam0=vmax_cam0,
-                vmin_cam=vmin_cam,
-                vmax_cam=vmax_cam,
-                vmin_cam_lamb=vmin_cam_lamb,
-                vmax_cam_lamb=vmax_cam_lamb,
-                vmin_plane=vmin_plane,
-                vmax_plane=vmax_plane,
-                # dout
-                **dout,
-            )
+        _plot(
+            coll=coll,
+            # extra
+            indplot=indplot,
+            dax=dax,
+            plot_config=plot_config,
+            fs=fs,
+            dmargin=dmargin,
+            dvminmax=dvminmax,
+            # dout
+            **dout,
+        )
 
     return dout
 
@@ -300,13 +276,15 @@ def _check(
     DR=None,
     DZ=None,
     Dphi=None,
+    # solid angle
+    config=None,
+    visibility=None,
     # bool
     verb=None,
     plot=None,
     # plotting
     indplot=None,
     plot_config=None,
-    config=None,
 ):
 
     # ----------
@@ -475,7 +453,7 @@ def _check(
     if any(lc):
 
         # key_mesh
-        wm = coll._whcih_mesh
+        wm = coll._which_mesh
         lok = list(coll.dobj.get(wm, {}).keys())
         key_mesh = ds._generic_check._check_var(
             key_mesh, 'key_mesh',
@@ -518,6 +496,16 @@ def _check(
         types=(int, float),
         default=defseg,
     ))
+
+    # -----------
+    # visibility
+    # -----------
+
+    visibility = ds._generic_check._check_var(
+        visibility, 'visibility',
+        types=bool,
+        default=vect is None,
+    )
 
     # -----------
     # verb
@@ -575,6 +563,7 @@ def _check(
         lop_pre, lop_post,
         res, margin_par, margin_perp,
         vect, segment, phi, Z, key_mesh,
+        visibility,
         verb, plot,
         indref, indplot,
         plot_config,
@@ -817,1170 +806,3 @@ def _get_intersect_los_plane(
     x1 = dx * e1[0] + dy * e1[1] + dz * e1[2]
 
     return x0, x1, iok
-
-
-# ###############################################################
-# ###############################################################
-#                   non-spectro
-# ###############################################################
-
-
-def _nonspectro(
-    coll=None,
-    key_cam=None,
-    doptics=None,
-    par=None,
-    is2d=None,
-    # pts
-    ptsx=None,
-    ptsy=None,
-    ptsz=None,
-    config=None,
-):
-
-    # -----------------
-    # prepare apertures
-
-    pinhole = doptics['pinhole']
-    if pinhole is False:
-        paths = doptics['paths']
-
-    apertures = coll.get_optics_as_input_solid_angle(doptics['optics'])
-
-    # -----------
-    # prepare det
-
-    k0, k1 = coll.dobj['camera'][key_cam]['dgeom']['outline']
-    cx, cy, cz = coll.get_camera_cents_xyz(key=key_cam)
-    dvect = coll.get_camera_unit_vectors(key=key_cam)
-
-    # -------------
-    # compute
-    # -------------
-
-    ref = (
-        coll.dobj['camera'][key_cam]['dgeom']['ref']
-        + tuple([None for ii in ptsx.shape])
-    )
-
-    # --------
-    # pinhole
-
-    if pinhole is True:
-
-        det = {
-            'cents_x': cx,
-            'cents_y': cy,
-            'cents_z': cz,
-            'outline_x0': coll.ddata[k0]['data'],
-            'outline_x1': coll.ddata[k1]['data'],
-            'nin_x': np.full(cx.shape, dvect['nin_x']) if par else dvect['nin_x'],
-            'nin_y': np.full(cx.shape, dvect['nin_y']) if par else dvect['nin_y'],
-            'nin_z': np.full(cx.shape, dvect['nin_z']) if par else dvect['nin_z'],
-            'e0_x': np.full(cx.shape, dvect['e0_x']) if par else dvect['e0_x'],
-            'e0_y': np.full(cx.shape, dvect['e0_y']) if par else dvect['e0_y'],
-            'e0_z': np.full(cx.shape, dvect['e0_z']) if par else dvect['e0_z'],
-            'e1_x': np.full(cx.shape, dvect['e1_x']) if par else dvect['e1_x'],
-            'e1_y': np.full(cx.shape, dvect['e1_y']) if par else dvect['e1_y'],
-            'e1_z': np.full(cx.shape, dvect['e1_z']) if par else dvect['e1_z'],
-        }
-
-        sang = _comp_solidangles.calc_solidangle_apertures(
-            # observation points
-            pts_x=ptsx,
-            pts_y=ptsy,
-            pts_z=ptsz,
-            # polygons
-            apertures=apertures,
-            detectors=det,
-            # possible obstacles
-            config=config,
-            # parameters
-            summed=False,
-            visibility=False,
-            return_vector=False,
-            return_flat_pts=False,
-            return_flat_det=False,
-        )
-
-    # -----------
-    # collimator
-
-    else:
-
-        sang = np.full(cx.shape + ptsx.shape, np.nan)
-        for ii, indch in enumerate(np.ndindex(cx.shape)):
-
-            det = {
-                'cents_x': cx[indch],
-                'cents_y': cy[indch],
-                'cents_z': cz[indch],
-                'outline_x0': coll.ddata[k0]['data'],
-                'outline_x1': coll.ddata[k1]['data'],
-                'nin_x': dvect['nin_x'] if par else dvect['nin_x'][indch],
-                'nin_y': dvect['nin_y'] if par else dvect['nin_y'][indch],
-                'nin_z': dvect['nin_z'] if par else dvect['nin_z'][indch],
-                'e0_x': dvect['e0_x'] if par else dvect['e0_x'][indch],
-                'e0_y': dvect['e0_y'] if par else dvect['e0_y'][indch],
-                'e0_z': dvect['e0_z'] if par else dvect['e0_z'][indch],
-                'e1_x': dvect['e1_x'] if par else dvect['e1_x'][indch],
-                'e1_y': dvect['e1_y'] if par else dvect['e1_y'][indch],
-                'e1_z': dvect['e1_z'] if par else dvect['e1_z'][indch],
-            }
-
-            sliap = indch + (slice(None),)
-            lap = [doptics['optics'][ii] for ii in paths[sliap].nonzero()[0]]
-            api = {kap: apertures[kap] for kap in lap}
-
-            sli = (indch, slice(None), slice(None))
-
-            sang[sli] = _comp_solidangles.calc_solidangle_apertures(
-                # observation points
-                pts_x=ptsx,
-                pts_y=ptsy,
-                pts_z=ptsz,
-                # polygons
-                apertures=api,
-                detectors=det,
-                # possible obstacles
-                config=config,
-                # parameters
-                summed=False,
-                visibility=False,
-                return_vector=False,
-                return_flat_pts=False,
-                return_flat_det=False,
-            )
-
-    axis = tuple([ii for ii in range(len(sang.shape[:-2]))])
-    return {
-        'sang0': {
-            'data': sang,
-            'ref': ref,
-            'units': 'sr',
-        },
-        'ndet': {
-            'data': np.sum(sang > 0., axis=axis),
-            'units': '',
-            'ref': ref[-2:],
-        },
-    }
-
-
-# ###############################################################
-# ###############################################################
-#                   Spectro
-# ###############################################################
-
-
-def _spectro(
-    coll=None,
-    key_diag=None,
-    key_cam=None,
-    is2d=None,
-    # pts
-    ptsx=None,
-    ptsy=None,
-    ptsz=None,
-    # solid angle
-    n0=None,
-    n1=None,
-    # res
-    res_lamb=None,
-):
-
-    # ------------
-    # prepare
-
-    dout = coll.get_raytracing_from_pts(
-        key=key_diag,
-        key_cam=key_cam,
-        # mesh
-        key_mesh=None,
-        res_RZ=None,
-        res_phi=None,
-        # points
-        ptsx=ptsx,
-        ptsy=ptsy,
-        ptsz=ptsz,
-        # solid angle
-        n0=n0,
-        n1=n1,
-        # others
-        res_rock_curve=None,
-        lamb0=None,
-        res_lamb=res_lamb,
-        plot=False,
-        plot_pixels=None,
-        plot_config=None,
-        vmin=None,
-        vmax=None,
-        append=False,
-    )
-
-    dout['sang0'] = {
-        'data': dout['sang0'],
-        'units': 'sr',
-    }
-    dout['sang'] = {
-        'data': dout['sang'],
-        'units': 'sr',
-    }
-    dout['sang_lamb'] = {
-        'data': dout['sang_lamb'],
-        'units': 'sr.m',
-    }
-
-    return dout
-
-
-# ###############################################################
-# ###############################################################
-#                   Plot from LOS
-# ###############################################################
-
-
-def _plot_from_los(
-    coll=None,
-    key_diag=None,
-    key_cam=None,
-    # dout
-    is2d=None,
-    spectro=None,
-    indch=None,
-    indref=None,
-    los_ref=None,
-    pt_ref=None,
-    klos=None,
-    e0=None,
-    e1=None,
-    ptsx=None,
-    ptsy=None,
-    ptsz=None,
-    x0=None,
-    x1=None,
-    dS=None,
-    vect=None,
-    sang0=None,
-    sang=None,
-    sang_lamb=None,
-    # extra
-    indplot=None,
-    dax=None,
-    plot_config=None,
-    fs=None,
-    dmargin=None,
-    vmin_cam0=None,
-    vmax_cam0=None,
-    vmin_cam=None,
-    vmax_cam=None,
-    vmin_cam_lamb=None,
-    vmax_cam_lamb=None,
-    vmin_plane=None,
-    vmax_plane=None,
-    **kwdargs,
-):
-
-    # ------------
-    # prepare
-
-    (
-        etend0, etend, etend_lamb,
-        etend_plane0, etend_plane, etend_plane_lamb,
-        sli, axis,
-        extent_cam, extent_plane,
-        vmin_cam0, vmax_cam0,
-        vmin_cam, vmax_cam,
-        vmin_cam_lamb, vmax_cam_lamb,
-        vmin_plane0, vmax_plane0,
-        vmin_plane, vmax_plane,
-        los_refr, los_refz,
-    ) = _check_plot(
-        coll=coll,
-        key_diag=key_diag,
-        key_cam=key_cam,
-        is2d=is2d,
-        spectro=spectro,
-        sang0=sang0,
-        sang=sang,
-        sang_lamb=sang_lamb,
-        x0=x0,
-        x1=x1,
-        dS=dS,
-        pt_ref=pt_ref,
-        los_ref=los_ref,
-        indref=indref,
-        indplot=indplot,
-        vmin_cam=vmin_cam,
-        vmax_cam=vmax_cam,
-        vmin_cam0=vmin_cam0,
-        vmax_cam0=vmax_cam0,
-        vmin_cam_lamb=vmin_cam_lamb,
-        vmax_cam_lamb=vmax_cam_lamb,
-        vmin_plane=vmin_plane,
-        vmax_plane=vmax_plane,
-    )
-
-    # --------------
-    # prepare
-    # --------------
-
-    ix1 = int(x1.size/2)
-    # sang_plot = sang0['data'][sli].ravel()
-    sang_tot_plot = np.sum(sang0['data'], axis=axis)
-
-    # tit
-    tit = (
-        "Solid angle subtended from points in the plasma by:\n"
-        f"{key_diag} - {key_cam}"
-    )
-
-    # --------------
-    # prepare
-
-    dplot = coll.get_diagnostic_dplot(
-        key=key_diag,
-        key_cam=key_cam,
-        optics=None,
-        elements='o',
-    )
-
-    # --------------
-    # pepare dax
-
-    if dax is None:
-        dax = _get_dax(
-            is2d=is2d,
-            spectro=spectro,
-            fs=fs,
-            dmargin=dmargin,
-            vect=vect,
-        )
-
-    dax = _generic_check._check_dax(dax=dax, main='cross')
-
-    # ----------
-    # points
-
-    kax = 'cross'
-    if dax.get(kax) is not None:
-        ax = dax[kax]['handle']
-
-        # pts
-        ax.scatter(
-            np.hypot(ptsx, ptsy).ravel(),
-            ptsz.ravel(),
-            c=sang_tot_plot,
-            s=4,
-            marker='.',
-            vmin=vmin_plane,
-            vmax=vmax_plane,
-        )
-
-        # los_ref
-        ax.plot(
-            los_refr,
-            los_refz,
-            c='k',
-            ls='-',
-            lw=1.,
-        )
-
-    kax = 'hor'
-    if dax.get(kax) is not None:
-        ax = dax[kax]['handle']
-
-        # pts
-        ax.scatter(
-            ptsx.ravel(),
-            ptsy.ravel(),
-            c=sang_tot_plot,
-            s=4,
-            marker='.',
-            vmin=vmin_plane,
-            vmax=vmax_plane,
-        )
-
-        # los_ref
-        ax.plot(
-            pt_ref[0] + np.r_[0, los_ref[0]],
-            pt_ref[1] + np.r_[0, los_ref[1]],
-            c='k',
-            ls='-',
-            lw=1.,
-        )
-
-    kax = '3d'
-    if dax.get(kax) is not None:
-        ax = dax[kax]['handle']
-
-        # pts
-        ax.scatter(
-            ptsx.ravel(),
-            ptsy.ravel(),
-            ptsz.ravel(),
-            c=sang_tot_plot,
-            s=4,
-            marker='.',
-            vmin=vmin_plane,
-            vmax=vmax_plane,
-        )
-
-        # los_ref
-        ax.plot(
-            pt_ref[0] + np.r_[0, los_ref[0]],
-            pt_ref[1] + np.r_[0, los_ref[1]],
-            pt_ref[2] + np.r_[0, los_ref[2]],
-            c='k',
-            ls='-',
-            lw=1.,
-        )
-
-    # ------------------
-    # integral per pixel
-
-    kax = 'cam_plane0'
-    if dax.get(kax) is not None:
-        ax = dax[kax]['handle']
-
-        if is2d is True:
-
-            # plane-specific etendue
-            im = ax.imshow(
-                etend_plane0.T,
-                extent=extent_cam,
-                origin='lower',
-                interpolation='nearest',
-                vmin=vmin_cam0,
-                vmax=vmax_cam0,
-            )
-
-            _add_marker(ax, indref, indplot)
-
-        else:
-            # per-pixel etendue
-            ax.plot(
-                etend0,
-                ls='-',
-                lw=1.,
-                c='k',
-                marker='.',
-                label='Analytical',
-            )
-
-            # plane-specific etendue
-            ax.plot(
-                etend_plane0,
-                ls='-',
-                lw=1.,
-                c='b',
-                marker='.',
-                label='Numerical',
-            )
-
-            # ref pixel
-            ax.axvline(
-                indref,
-                c='k',
-                ls='--',
-                lw=1,
-            )
-
-            # plot pixel
-            ax.axvline(
-                indplot,
-                c='g',
-                ls='--',
-                lw=1,
-            )
-
-            ax.legend()
-
-    kax = 'cam_etend0'
-    if dax.get(kax) is not None:
-        ax = dax[kax]['handle']
-
-        if is2d is True:
-
-            # plane-specific etendue
-            im0 = ax.imshow(
-                etend0.T,
-                extent=extent_cam,
-                origin='lower',
-                interpolation='nearest',
-                vmin=vmin_cam0,
-                vmax=vmax_cam0,
-            )
-
-            _add_marker(ax, indref, indplot)
-            plt.colorbar(im, ax=[ax, dax['cam_plane0']['handle']])
-
-    kax = 'cam_diff0'
-    if dax.get(kax) is not None:
-        ax = dax[kax]['handle']
-
-        if is2d is True:
-
-            # plane-specific etendue
-            diff = (etend_plane0 - etend0).T
-            err = np.nanmax(np.abs(diff))
-            imdiff = ax.imshow(
-                diff,
-                extent=extent_cam,
-                origin='lower',
-                interpolation='nearest',
-                vmin=-err,
-                vmax=err,
-                cmap=plt.cm.seismic,
-            )
-
-            # ref pixel
-            _add_marker(ax, indref, indplot)
-            plt.colorbar(imdiff, ax=ax)
-
-    # -------------------------------------
-    # integral per pixel with rocking curve
-
-    kax = 'cam_plane'
-    if dax.get(kax) is not None:
-        ax = dax[kax]['handle']
-
-        # plane-specific etendue
-        im = ax.imshow(
-            etend_plane.T,
-            extent=extent_cam,
-            origin='lower',
-            interpolation='nearest',
-            vmin=vmin_cam,
-            vmax=vmax_cam,
-        )
-
-        # ref pixel
-        _add_marker(ax, indref, indplot)
-
-    kax = 'cam_etend'
-    if dax.get(kax) is not None:
-        ax = dax[kax]['handle']
-
-        # plane-specific etendue
-        im0 = ax.imshow(
-            etend.T,
-            extent=extent_cam,
-            origin='lower',
-            interpolation='nearest',
-            vmin=vmin_cam,
-            vmax=vmax_cam,
-        )
-
-        # ref pixel
-        _add_marker(ax, indref, indplot)
-        plt.colorbar(im, ax=[ax, dax['cam_plane']['handle']])
-
-    kax = 'cam_diff'
-    if dax.get(kax) is not None:
-        ax = dax[kax]['handle']
-
-        # plane-specific etendue
-        diff = (etend_plane - etend).T
-        err = np.nanmax(np.abs(diff))
-        imdiff = ax.imshow(
-            diff,
-            extent=extent_cam,
-            origin='lower',
-            interpolation='nearest',
-            vmin=-err,
-            vmax=err,
-            cmap=plt.cm.seismic,
-        )
-
-        # ref pixel
-        _add_marker(ax, indref, indplot)
-        plt.colorbar(imdiff, ax=ax)
-
-    # -----------------------------------------------
-    # integral per pixel with rocking curve and dlamb
-
-    kax = 'cam_plane_lamb'
-    if dax.get(kax) is not None:
-        ax = dax[kax]['handle']
-
-        # plane-specific etendue
-        im = ax.imshow(
-            etend_plane_lamb.T,
-            extent=extent_cam,
-            origin='lower',
-            interpolation='nearest',
-            vmin=vmin_cam_lamb,
-            vmax=vmax_cam_lamb,
-        )
-
-        # ref pixel
-        _add_marker(ax, indref, indplot)
-
-    kax = 'cam_etend_lamb'
-    if dax.get(kax) is not None:
-        ax = dax[kax]['handle']
-
-        # plane-specific etendue
-        im0 = ax.imshow(
-            etend_lamb.T,
-            extent=extent_cam,
-            origin='lower',
-            interpolation='nearest',
-            vmin=vmin_cam_lamb,
-            vmax=vmax_cam_lamb,
-        )
-
-        # ref pixel
-        _add_marker(ax, indref, indplot)
-        plt.colorbar(im0, ax=[ax, dax['cam_plane_lamb']['handle']])
-
-    kax = 'cam_diff_lamb'
-    if dax.get(kax) is not None:
-        ax = dax[kax]['handle']
-
-        # plane-specific etendue
-        diff = (etend_plane_lamb - etend_lamb).T
-        err = np.nanmax(np.abs(diff))
-        imdiff = ax.imshow(
-            diff,
-            extent=extent_cam,
-            origin='lower',
-            interpolation='nearest',
-            vmin=-err,
-            vmax=err,
-            cmap=plt.cm.seismic,
-        )
-
-        # ref pixel
-        _add_marker(ax, indref, indplot)
-        plt.colorbar(imdiff, ax=ax)
-
-    # -------------------------------------------
-    # sang per plane point for selected pixel
-
-    kax = 'plane0'
-    if dax.get(kax) is not None:
-        ax = dax[kax]['handle']
-
-        im = ax.imshow(
-            sang_tot_plot.T,
-            # sang0['data'][sli].T,
-            extent=extent_plane,
-            origin='lower',
-            interpolation='nearest',
-            vmin=vmin_plane0,
-            vmax=vmax_plane0,
-        )
-
-        ax.axhline(x1[ix1], c='w', ls='--', lw=0.5)
-
-        plt.colorbar(
-            im,
-            ax=ax,
-            cax=dax.get('colorbar', {}).get('handle'),
-            label='solid angle (sr)',
-        )
-
-    kax = 'plane'
-    if dax.get(kax) is not None:
-        ax = dax[kax]['handle']
-
-        im = ax.imshow(
-            sang['data'][sli].T,
-            extent=extent_plane,
-            origin='lower',
-            interpolation='nearest',
-            vmin=vmin_plane,
-            vmax=vmax_plane,
-        )
-
-        plt.colorbar(im, ax=ax, label='solid angle (sr)')
-
-    # -------------------------------------------
-    # slice through sang 2d map
-
-    kax = 'slice'
-    if dax.get(kax) is not None:
-        ax = dax[kax]['handle']
-
-        ax.plot(
-            x0,
-            sang0['data'][..., ix1].T.reshape((x0.size, -1)),
-        )
-
-    # ----------
-    # diag geom
-
-    _class8_plot._plot_diag_geom(
-        dax=dax,
-        key_cam=key_cam,
-        dplot=dplot,
-        is2d=coll.dobj['diagnostic'][key_diag]['is2d'],
-    )
-
-    # -------
-    # config
-
-    if plot_config.__class__.__name__ == 'Config':
-
-        kax = 'cross'
-        if dax.get(kax) is not None:
-            ax = dax[kax]['handle']
-            plot_config.plot(lax=ax, proj=kax, dLeg=False)
-
-        kax = 'hor'
-        if dax.get(kax) is not None:
-            ax = dax[kax]['handle']
-            plot_config.plot(lax=ax, proj=kax, dLeg=False)
-
-    # ------------
-    # figure title
-
-    list(dax.values())[0]['handle'].figure.suptitle(
-        tit,
-        size=14,
-        fontweight='bold',
-    )
-
-    return dax
-
-
-def _check_plot(
-    coll=None,
-    key_diag=None,
-    key_cam=None,
-    is2d=None,
-    spectro=None,
-    sang0=None,
-    sang=None,
-    sang_lamb=None,
-    x0=None,
-    x1=None,
-    dS=None,
-    pt_ref=None,
-    los_ref=None,
-    indref=None,
-    indplot=None,
-    vmin_cam=None,
-    vmax_cam=None,
-    vmin_cam0=None,
-    vmax_cam0=None,
-    vmin_cam_lamb=None,
-    vmax_cam_lamb=None,
-    vmin_plane=None,
-    vmax_plane=None,
-    vmin_plane0=None,
-    vmax_plane0=None,
-):
-
-    # ---------
-    # prepare
-    # ---------
-
-    doptics = coll.dobj['diagnostic'][key_diag]['doptics'][key_cam]
-
-    # ---------
-    # data
-    # ---------
-
-    if is2d is True:
-        # sang['data']
-        pass
-    else:
-        pass
-
-    # ---------
-    # integral
-    # ---------
-
-    if spectro:
-        etend_plane0 = np.nansum(
-            np.nansum(sang0['data'], axis=-1),
-            axis=-1
-        ) * dS
-        etend_plane = np.nansum(
-            np.nansum(sang['data'], axis=-1),
-            axis=-1,
-        ) * dS
-
-        ketend = doptics['etendue0']
-        etend0 = coll.ddata[ketend]['data']
-        ketend = doptics['etendue']
-        etend = coll.ddata[ketend]['data']
-
-        dlamb = coll.get_diagnostic_data(
-            key_diag,
-            key_cam=key_cam,
-            data='dlamb',
-        )[0][key_cam]
-
-        etend_lamb = etend * dlamb
-        etend_plane_lamb = sang_lamb['data'] * dS
-
-    else:
-        etend_plane0 = np.nansum(
-            np.nansum(sang0['data'], axis=-1),
-            axis=-1,
-        ) * dS
-        etend_plane = None
-
-        ketend = coll.dobj['diagnostic'][key_diag]['doptics'][key_cam]['etendue']
-        etend0 = coll.ddata[ketend]['data']
-        etend = None
-
-        etend_lamb = None
-        etend_plane_lamb = None
-
-    # ---------
-    # extent
-
-    # cam
-    extent_cam = None
-    if is2d:
-        extent_cam = None
-        sli = (indplot[0], indplot[1], slice(None), slice(None))
-        axis = (0, 1)
-    else:
-        sli = (indplot, slice(None), slice(None))
-        axis = 0
-
-    # extent_plane
-    dx0 = 0.5*(x0[1] - x0[0])
-    dx1 = 0.5*(x1[1] - x1[0])
-    extent_plane = (
-        x0.min() - dx0,
-        x0.max() + dx0,
-        x1.min() - dx1,
-        x1.max() + dx1,
-    )
-
-    # ----------
-    # vmin, vmax
-
-    if vmin_cam0 is None:
-        vmin_cam0 = min(
-            np.nanmin(etend0), np.nanmin(etend_plane0)
-        )
-    if vmax_cam0 is None:
-        vmax_cam0 = max(
-            np.nanmax(etend0), np.nanmax(etend_plane0)
-        )
-
-    if spectro:
-        if vmin_cam is None:
-            vmin_cam = min(
-                np.nanmin(etend), np.nanmin(etend_plane)
-            )
-        if vmax_cam is None:
-            vmax_cam = max(
-                np.nanmax(etend), np.nanmax(etend_plane)
-            )
-        if vmin_cam_lamb is None:
-            vmin_cam_lamb = min(
-                np.nanmin(etend_lamb), np.nanmin(etend_plane_lamb)
-            )
-        if vmax_cam_lamb is None:
-            vmax_cam_lamb = max(
-                np.nanmax(etend_lamb), np.nanmax(etend_plane_lamb)
-            )
-
-    else:
-        vmin_cam, vmax_cam = None, None
-        vmin_cam_lamb, vmax_cam_lamb = None, None
-
-    if vmin_plane0 is None:
-        vmin_plane0 = np.nanmin(sang0['data'])
-    if vmax_plane0 is None:
-        vmax_plane0 = np.nanmax(sang0['data'])
-
-    if sang is not None:
-        if vmin_plane is None:
-            vmin_plane = np.nanmin(sang['data'])
-        if vmax_plane is None:
-            vmax_plane = np.nanmax(sang['data'])
-
-    # --------
-    # los_ref
-
-    add = np.linspace(0, 1, 20)
-    los_refr = np.hypot(
-        pt_ref[0] + add * los_ref[0],
-        pt_ref[1] + add * los_ref[1],
-    )
-    los_refz = pt_ref[2] + add * los_ref[2]
-
-    return (
-        etend0, etend, etend_lamb,
-        etend_plane0, etend_plane, etend_plane_lamb,
-        sli, axis,
-        extent_cam, extent_plane,
-        vmin_cam0, vmax_cam0,
-        vmin_cam, vmax_cam,
-        vmin_cam_lamb, vmax_cam_lamb,
-        vmin_plane0, vmax_plane0,
-        vmin_plane, vmax_plane,
-        los_refr, los_refz,
-    )
-
-
-def _get_dax(
-    is2d=None,
-    spectro=None,
-    fs=None,
-    dmargin=None,
-    vect=None,
-):
-
-    # ----------
-    # check
-
-    if fs is None:
-        fs = (16, 10)
-
-    dmargin = ds._generic_check._check_var(
-        dmargin, 'dmargin',
-        types=dict,
-        default={
-            'bottom': 0.05, 'top': 0.92,
-            'left': 0.05, 'right': 0.95,
-            'wspace': 0.4, 'hspace': 0.5,
-        },
-    )
-
-    # --------
-    # prepare
-
-    fig = plt.figure(figsize=fs)
-
-    if spectro is True:
-        gs = gridspec.GridSpec(ncols=8, nrows=8, **dmargin)
-
-        # --------
-        # create
-
-        # geometry
-        ax0 = fig.add_subplot(
-            gs[:2, :2],
-            aspect='equal',
-            adjustable='datalim',
-        )
-        ax0.set_xlabel('X (m)')
-        ax0.set_ylabel('Y (m)')
-
-        ax1 = fig.add_subplot(
-            gs[2:4, :2],
-            aspect='equal',
-            adjustable='datalim',
-        )
-        ax1.set_xlabel('R (m)')
-        ax1.set_ylabel('Z (m)')
-
-        ax2 = fig.add_subplot(gs[4:, :2], projection='3d')
-        ax2.set_xlabel('X (m)')
-        ax2.set_ylabel('Y (m)')
-        ax2.set_ylabel('Z (m)')
-
-        # images for etend0
-        ax30 = fig.add_subplot(gs[:2, 2:4], aspect='equal')
-        ax30.set_ylabel('x1 (m)')
-        ax30.set_xlabel('x0 (m)')
-        ax30.set_title(
-            'etendue\nw/o rock. curve',
-            size=12,
-            fontweight='bold',
-        )
-
-        ax31 = fig.add_subplot(gs[:2, 4:6], sharex=ax30, sharey=ax30)
-        ax31.set_xlabel('x0 (m)')
-        ax31.set_title(
-            'integral\nw/o rock. curve',
-            size=12,
-            fontweight='bold',
-        )
-
-        ax32 = fig.add_subplot(gs[:2, 6:], sharex=ax30, sharey=ax30)
-        ax32.set_xlabel('x0 (m)')
-        ax32.set_title('difference', size=12, fontweight='bold')
-
-        # images for etend
-        ax40 = fig.add_subplot(gs[2:4, 2:4], aspect='equal')
-        ax40.set_ylabel('x1 (m)')
-        ax40.set_xlabel('x0 (m)')
-        ax40.set_title(
-            'etendue\nwith rock. curve',
-            size=12,
-            fontweight='bold',
-        )
-
-        ax41 = fig.add_subplot(gs[2:4, 4:6], sharex=ax30, sharey=ax30)
-        ax41.set_xlabel('x0 (m)')
-        ax41.set_title(
-            'integral\nwith rock. curve',
-            size=12,
-            fontweight='bold',
-        )
-
-        ax42 = fig.add_subplot(gs[2:4, 6:], sharex=ax30, sharey=ax30)
-        ax42.set_xlabel('x0 (m)')
-        ax42.set_title('difference', size=12, fontweight='bold')
-
-        # images for etend * dlamb
-        ax50 = fig.add_subplot(gs[4:6, 2:4], aspect='equal')
-        ax50.set_ylabel('x1 (m)')
-        ax50.set_xlabel('x0 (m)')
-        ax50.set_title(
-            'etendue * dlamb\nwith rock. curve',
-            size=12,
-            fontweight='bold',
-        )
-
-        ax51 = fig.add_subplot(gs[4:6, 4:6], sharex=ax30, sharey=ax30)
-        ax51.set_xlabel('x0 (m)')
-        ax51.set_title(
-            'integral * dlamb\nwith rock. curve',
-            size=12,
-            fontweight='bold',
-        )
-
-        ax52 = fig.add_subplot(gs[4:6, 6:], sharex=ax30, sharey=ax30)
-        ax52.set_xlabel('x0 (m)')
-        ax52.set_title('difference', size=12, fontweight='bold')
-
-        # plane
-        ax60 = fig.add_subplot(
-            gs[-2:, 2:5],
-            aspect='equal',
-            adjustable='datalim',
-        )
-        ax60.set_ylabel('x1 (m)')
-        ax60.set_xlabel('x0 (m)')
-
-        ax61 = fig.add_subplot(
-            gs[-2:, 5:],
-            aspect='equal',
-            adjustable='datalim',
-        )
-        ax61.set_ylabel('x1 (m)')
-        ax61.set_xlabel('x0 (m)')
-
-        # dict
-        dax = {
-            'cross': {'handle': ax0, 'type': 'cross'},
-            'hor': {'handle': ax1, 'type': 'hor'},
-            '3d': {'handle': ax2, 'type': '3d'},
-            'cam_etend0': {'handle': ax30, 'type': 'camera'},
-            'cam_plane0': {'handle': ax31, 'type': 'camera'},
-            'cam_diff0': {'handle': ax32, 'type': 'camera'},
-            'cam_etend': {'handle': ax40, 'type': 'camera'},
-            'cam_plane': {'handle': ax41, 'type': 'camera'},
-            'cam_diff': {'handle': ax42, 'type': 'camera'},
-            'cam_etend_lamb': {'handle': ax50, 'type': 'camera'},
-            'cam_plane_lamb': {'handle': ax51, 'type': 'camera'},
-            'cam_diff_lamb': {'handle': ax52, 'type': 'camera'},
-            'plane0': {'handle': ax60, 'type': 'misc'},
-            'plane': {'handle': ax61, 'type': 'misc'},
-        }
-
-    else:
-        nn = 4
-        gs = gridspec.GridSpec(ncols=5*nn+1, nrows=6, **dmargin)
-
-        # --------
-        # create
-
-        ax0 = fig.add_subplot(gs[:3, :2*nn], aspect='equal', adjustable='datalim')
-        ax0.set_xlabel('X (m)', size=12, fontweight='bold')
-        ax0.set_ylabel('Y (m)', size=12, fontweight='bold')
-
-        ax1 = fig.add_subplot(gs[3:, :2*nn], aspect='equal', adjustable='datalim')
-        ax1.set_xlabel('R (m)', size=12, fontweight='bold')
-        ax1.set_ylabel('Z (m)', size=12, fontweight='bold')
-
-        ax2 = fig.add_subplot(gs[:2, 2*nn+1:3*nn+1], projection='3d')
-        ax2.set_xlabel('X (m)', size=12, fontweight='bold')
-        ax2.set_ylabel('Y (m)', size=12, fontweight='bold')
-        ax2.set_ylabel('Z (m)', size=12, fontweight='bold')
-
-        # etendue
-        if vect == 'nin':
-            if is2d is True:
-                ax30 = fig.add_subplot(gs[2:4, 2*nn:3*nn], aspect='equal')
-                ax30.set_ylabel('x1 (m)')
-                ax30.set_xlabel('x0 (m)')
-                ax30.set_title('integral', size=12, fontweight='bold')
-
-                ax31 = fig.add_subplot(
-                    gs[2:4, 3*nn:4*nn],
-                    sharex=ax30,
-                    sharey=ax30,
-                )
-                ax31.set_xlabel('x0 (m)')
-                ax31.set_title('etendue', size=12, fontweight='bold')
-
-                ax32 = fig.add_subplot(
-                    gs[2:4, 4*nn:-1],
-                    sharex=ax30,
-                    sharey=ax30,
-                )
-                ax32.set_xlabel('x0 (m)')
-                ax32.set_title('difference', size=12, fontweight='bold')
-
-            else:
-                ax30 = fig.add_subplot(gs[:2, 3*nn+2:])
-                ax30.set_xlabel('channel', size=12, fontweight='bold')
-                ax30.set_ylabel('Etendue (m2.sr)', size=12, fontweight='bold')
-
-        ax4 = fig.add_subplot(
-            gs[2:4, 2*nn+1:-1],
-            aspect='equal',
-            adjustable='datalim',
-        )
-        ax4.set_ylabel('x1 (m)', size=12, fontweight='bold')
-        ax4.set_xlabel('x0 (m)', size=12, fontweight='bold')
-
-        ax4c = fig.add_subplot(gs[2:4, -1])
-
-        ax5 = fig.add_subplot(gs[4:, 2*nn+1:-1], sharex=ax4)
-        ax5.set_ylabel('sang (sr)', size=12, fontweight='bold')
-        ax5.set_xlabel('x0 (m)', size=12, fontweight='bold')
-
-        # dict
-        dax = {
-            'cross': {'handle': ax0, 'type': 'cross'},
-            'hor': {'handle': ax1, 'type': 'hor'},
-            '3d': {'handle': ax2, 'type': '3d'},
-            'plane0': {'handle': ax4, 'type': 'misc'},
-            'colorbar': {'handle': ax4c, 'type': 'misc'},
-            'slice': {'handle': ax5, 'type': 'misc'},
-        }
-        if vect == 'nin':
-            dax['cam_plane0'] = {'handle': ax30, 'type': 'camera'}
-            if is2d is True:
-                dax['cam_etend0'] = {'handle': ax31, 'type': 'camera'}
-                dax['cam_diff0'] = {'handle': ax32, 'type': 'camera'}
-
-    return dax  # , cax
-
-
-def _add_marker(ax=None, indref=None, indplot=None):
-
-    # ref pixel
-    ax.plot(
-        [indref[0]],
-        [indref[1]],
-        marker='s',
-        markerfacecolor='None',
-        markeredgecolor='k',
-        ms=4,
-    )
-
-    # plot pixel
-    ax.plot(
-        [indplot[0]],
-        [indplot[1]],
-        marker='s',
-        markerfacecolor='None',
-        markeredgecolor='g',
-        ms=4,
-    )
